requirements.txt
flet==0.24.1
Pyrebase4==4.8.0